+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Please mark all change in change log and use the ticket from JIRA.
 ## Feature
 
 ## Improvement
+- \#204 - improve grpc performance in search
 - \#207 - Add more unittest for config set/get
 - \#208 - optimize unittest to support run single test more easily